refactor: separate realtimechat folder for services
